refactor(cms): split homepage content helpers
Move homepage normalization and validation into cms/homepage.ts so cms/index.ts stays focused on fetch/query orchestration and public re-exports.

Verified with pnpm --filter @anipotts/lib typecheck; pnpm --filter @anipotts/lib test; pnpm --filter @anipotts/lib build; pnpm dlx knip --no-exit-code; pnpm exec prettier --check packages/lib/src/cms/index.ts packages/lib/src/cms/homepage.ts; git diff --check.
